Analysis constraint error should not be closeable (#5572)
* Analysis constraint error should not be closeable

When an error occurs because for example a variable does not have enough levels, this error should not be closeable: the user must remove it from the VariablesList first. As long such an error exists, the analysis cannot be run.

* Don't refresh analysis if it has an error

+++ Desktop/analysis/analysis.cpp
@@ -222,6 +222,9 @@ void Analysis::run()
 
 void Analysis::refresh()
 {
+	if (form() && form()->hasError())
+		return;
+
 	TempFiles::deleteAll(int(_id));
 	run();
